02-gitlab-ci/example/src/pyproject.toml
[tool.poetry]
name = "share_service"
version = "0.1.0"
description = ""
authors = ["Ruslan Mamleev <ruslan@mircod.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
psycopg2-binary = "^2.8.6"
django-stdimage = "^5.1.1"
celery = "4.3.0"
flower = "^0.9.7"
redis = "^3.5.3"
django-rest-framework = "^0.1.0"
drf-yasg = "^1.20.0"
django-rest-framework-social-oauth2 = "^1.1.0"
pytest = "^6.2.2"
coverage = "^5.4"
factory-boy = "^3.2.0"
pytest-django = "^4.1.0"
gunicorn = "^20.0.4"
django-3-jet = "^1.0.8"
django-celery-beat = "1.4.0"
django-crum = "^0.7.9"
sentry-sdk = "^1.0.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
isort = "^5.7.0"
black = "19.10b0"
flake8 = "^3.8.4"
flake8-polyfill = "^1.0.2"
radon = "^4.3.2"
django-rosetta = "^0.9.5"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
